move keyboard related code to keyboard.hpp
