+++ MatrixGame/src/MatrixTerSurface.cpp
@@ -12,6 +12,8 @@
 #include "CFile.hpp"
 #include "Mem.hpp"
 
+#include <keyboard.hpp>
+
 CBigIB *CTerSurface::m_BigIB;
 CBigVB<CTerSurface::STerSurfVertex> *CTerSurface::m_BigVB;
 CBigVB<CTerSurface::STerSurfVertexM> *CTerSurface::m_BigVBM;
